+++ src/Contracts/Changelog.md
@@ -8,7 +8,7 @@
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support
 - Added the `LineItemType` interface
-- Added the konekt/enum package dependency (v4.2)
+- Added the konekt/xtend package dependency (v2.0)
 
 ## 5.x Series